use wait_until_true to simplify the test logic
